Moved all of the API to core.
