+++ nodewatcher/modules/authentication/public_key/resources.py
@@ -1,6 +1,6 @@
 from tastypie import authorization as api_authorization, authentication as api_authentication
 
-from nodewatcher.core.frontend import api
+from nodewatcher.core import api
 
 from . import models
