+++ nodewatcher/modules/frontend/generator/frontend.py
@@ -1,7 +1,8 @@
 from django.conf import urls
 from django.core import urlresolvers
 
-from nodewatcher.core.frontend import api, components
+from nodewatcher.core import api
+from nodewatcher.core.frontend import components
 
 from . import resources, views
