+++ nodewatcher/modules/frontend/list/frontend.py
@@ -1,6 +1,7 @@
 from django.core import urlresolvers
 
-from nodewatcher.core.frontend import api, components
+from nodewatcher.core import api
+from nodewatcher.core.frontend import components
 
 from . import resources, views
