+++ nodewatcher/modules/routing/olsr/resources.py
@@ -1,8 +1,8 @@
 from django.db import models as django_models
 
 from nodewatcher.core import models as core_models, resources as core_resources
-from nodewatcher.core.frontend import api
-from nodewatcher.core.frontend.api import fields as api_fields
+from nodewatcher.core import api
+from nodewatcher.core.api import fields as api_fields
 # TODO: Remove this dependency after we have https://dev.wlan-si.net/ticket/1268.
 from nodewatcher.modules.frontend.list import resources
